Add regex that allows a job to only be executed for specific branches
I customized the GitlabBuildTrigger and added a new field in which the user can specify a regex.
This regex is used in the GitlabMergeRequestWrapper to decide whether the job should be executed or not.
The target branch name will not be checked using the regex when no regex has been defined in the trigger field.